# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MindBridge psychiatric care platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Provider {
  id: number;
  name: string;
  title: string;
  specialty: string;
  bio: string;
  rating: number;
  reviewCount: number;
  yearsExperience: number;
  imageUrl: string;
  available: boolean;
  sessionPrice: number;
  /** @nullable */
  languages?: string[] | null;
  acceptsInsurance?: boolean;
  /** @nullable */
  nextAvailable?: string | null;
}

export interface ProviderInput {
  name: string;
  title: string;
  specialty: string;
  bio: string;
  yearsExperience: number;
  imageUrl: string;
  sessionPrice: number;
  languages?: string[];
  acceptsInsurance?: boolean;
}

export interface Appointment {
  id: number;
  patientName: string;
  patientEmail: string;
  providerId: number;
  providerName: string;
  date: string;
  time: string;
  type: string;
  status: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  patientPhone?: string | null;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  careType?: string | null;
  /** @nullable */
  diagnosisSummary?: string | null;
  /** @nullable */
  approvedAt?: string | null;
  /** @nullable */
  approvedBy?: string | null;
  createdAt: string;
}

export interface AppointmentInput {
  patientName: string;
  patientEmail: string;
  providerId: number;
  date: string;
  time: string;
  type: string;
  notes?: string;
  patientPhone?: string;
  paymentMethod?: string;
  careType?: string;
  diagnosisSummary?: string;
}

export interface AppointmentUpdate {
  status?: string;
  notes?: string;
  approvedBy?: string;
}

export interface Specialty {
  id: number;
  name: string;
  description: string;
  iconName: string;
  providerCount: number;
}

export interface Testimonial {
  id: number;
  patientName: string;
  content: string;
  rating: number;
  condition: string;
  createdAt: string;
  avatarInitials?: string;
}

export interface TestimonialInput {
  patientName: string;
  content: string;
  rating: number;
  condition: string;
}

export interface PlatformStats {
  totalProviders: number;
  totalPatients: number;
  totalSessions: number;
  avgRating: number;
  countriesServed: number;
  satisfactionRate: number;
}

export interface Payment {
  id: number;
  /** @nullable */
  appointmentId?: number | null;
  patientName: string;
  /** @nullable */
  patientPhone?: string | null;
  /** @nullable */
  patientEmail?: string | null;
  providerName: string;
  /** @nullable */
  providerId?: number | null;
  amount: number;
  currency: string;
  method: string;
  status: string;
  /** @nullable */
  reference?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface PaymentInput {
  patientName: string;
  patientPhone?: string;
  patientEmail?: string;
  providerName: string;
  providerId?: number;
  appointmentId?: number;
  amount: number;
  currency?: string;
  method?: string;
  status?: string;
  reference?: string;
  notes?: string;
}

export interface Rating {
  id: number;
  /** @nullable */
  appointmentId?: number | null;
  providerId: number;
  patientName: string;
  rating: number;
  /** @nullable */
  mood?: string | null;
  /** @nullable */
  notes?: string | null;
  sessionDate: string;
  createdAt: string;
}

export interface RatingInput {
  appointmentId?: number;
  providerId: number;
  patientName: string;
  rating: number;
  mood?: string;
  notes?: string;
  sessionDate: string;
}

export interface PaymentStatusUpdate {
  status?: string;
  reference?: string;
  notes?: string;
}

export type AppointmentBulkUpdateStatus = typeof AppointmentBulkUpdateStatus[keyof typeof AppointmentBulkUpdateStatus];


export const AppointmentBulkUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
  completed: 'completed',
} as const;

export interface AppointmentBulkUpdate {
  ids: number[];
  status: AppointmentBulkUpdateStatus;
}

export type ListProvidersParams = {
specialty?: string;
available?: boolean;
search?: string;
};

export type GetBookedSlotsParams = {
providerId: number;
date: string;
};

export type GetBookedSlots200 = {
  bookedSlots?: string[];
};

export type BulkUpdateAppointments200 = {
  updated?: number;
  appointments?: Appointment[];
};