# lib/api-zod/src/generated/types/appointmentBulkUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MindBridge psychiatric care platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentBulkUpdateStatus } from './appointmentBulkUpdateStatus';

export interface AppointmentBulkUpdate {
  ids: number[];
  status: AppointmentBulkUpdateStatus;
}