# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MindBridge psychiatric care platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Payment {
  id: number;
  /** @nullable */
  appointmentId?: number | null;
  patientName: string;
  /** @nullable */
  patientPhone?: string | null;
  /** @nullable */
  patientEmail?: string | null;
  providerName: string;
  /** @nullable */
  providerId?: number | null;
  amount: number;
  currency: string;
  method: string;
  status: string;
  /** @nullable */
  reference?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}