# lib/api-zod/src/generated/types/paymentInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MindBridge psychiatric care platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentInput {
  patientName: string;
  patientPhone?: string;
  patientEmail?: string;
  providerName: string;
  providerId?: number;
  appointmentId?: number;
  amount: number;
  currency?: string;
  method?: string;
  status?: string;
  reference?: string;
  notes?: string;
}