# lib/api-zod/src/generated/types/paymentStatusUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MindBridge psychiatric care platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentStatusUpdate {
  status?: string;
  reference?: string;
  notes?: string;
}