# lib/api-zod/src/generated/types/rating.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MindBridge psychiatric care platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Rating {
  id: number;
  /** @nullable */
  appointmentId?: number | null;
  providerId: number;
  patientName: string;
  rating: number;
  /** @nullable */
  mood?: string | null;
  /** @nullable */
  notes?: string | null;
  sessionDate: string;
  createdAt: string;
}