# lib/api-zod/src/generated/types/ratingInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MindBridge psychiatric care platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RatingInput {
  appointmentId?: number;
  providerId: number;
  patientName: string;
  rating: number;
  mood?: string;
  notes?: string;
  sessionDate: string;
}