# lib/db/src/schema/ratings.ts
import { pgTable, serial, text, integer, timestamp, index } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const ratingsTable = pgTable("ratings", {
  id: serial("id").primaryKey(),
  appointmentId: integer("appointment_id"),
  providerId: integer("provider_id").notNull(),
  patientName: text("patient_name").notNull(),
  rating: integer("rating").notNull(),
  mood: text("mood"),
  notes: text("notes"),
  sessionDate: text("session_date").notNull(),
  createdAt: timestamp("created_at").notNull().defaultNow(),
}, (t) => [
  index("ratings_provider_id_idx").on(t.providerId),
]);

export const insertRatingSchema = createInsertSchema(ratingsTable).omit({ id: true, createdAt: true });
export type InsertRating = z.infer<typeof insertRatingSchema>;
export type Rating = typeof ratingsTable.$inferSelect;